+++ migrations/900_institutions__create.sql
@@ -1,0 +1,13 @@
+CREATE TABLE institutions (
+    id bigserial PRIMARY KEY,
+    institution TEXT NOT NULL,
+    uid_pattern TEXT
+);
+
+CREATE INDEX IF NOT EXISTS institutions_institution_idx ON institutions(institution);
+
+INSERT INTO institutions (institution, uid_pattern) VALUES
+    ('UIUC', '%@illinois.edu'),
+    ('ZJUI', '%@intl.zju.edu.cn'),
+    ('gvsu.edu', '%@mail.gvsu.edu'),
+    ('Any', '%');