+++ sync/fromDisk/courseInstances.js
@@ -1,6 +1,7 @@
 const _ = require('lodash');
 const { callbackify } = require('util');
 const sqldb = require('@prairielearn/prairielib/sql-db');
+const config = require('../../lib/config');
 
 module.exports.sync = function(courseInfo, courseInstanceDB, callback) {
     callbackify(async () => {
@@ -21,7 +22,7 @@ module.exports.sync = function(courseInfo, courseInstanceDB, callback) {
                 uids: _(accessRule).has('uids') ? accessRule.uids : null,
                 start_date: _(accessRule).has('startDate') ? accessRule.startDate : null,
                 end_date: _(accessRule).has('endDate') ? accessRule.endDate : null,
-                institution: _(accessRule).has('institution') ? accessRule.institution : 'UIUC',
+                institution: _(accessRule).has('institution') ? accessRule.institution : config.institutionDefault,
             }));
 
             return {
@@ -29,7 +30,7 @@ module.exports.sync = function(courseInfo, courseInstanceDB, callback) {
                 short_name: courseInstanceShortName,
                 long_name: courseInstance.longName,
                 number: courseInstance.number,
-                display_timezone: courseInstance.timezone || courseInfo.timezone || 'America/Chicago',
+                display_timezone: courseInstance.timezone || courseInfo.timezone || config.timezoneDefault,
                 access_rules: accessRules,
             };
         });