+++ ChangeLog.md
@@ -66,8 +66,10 @@
   * Change docker to use CentOS7 python3 instead of python36u (Dave Mussulman).
 
   * Change SSL file paths to be configurable (Dave Mussulman).
-  
+
   * Change `rpy2` version from 2.9.5 to 3.2.0 (James Balamuta).
+
+  * Change institution definitions to read from the `institutions` table (Dave Mussulman).
 
   * Fix dead letter cron job for `async` v3 (Matt West).
 
@@ -120,7 +122,7 @@
   * Fix git merge during CI (Matt West).
 
   * Fix to prevent instructor testing of externally-graded questions (Matt West).
-  
+
   * Fix LTI outcome reporting with Blackboard Learn (Dave Mussulman).
 
 * __3.2.0__ - 2019-08-05